Prediction using Supervised ML
Introduction
Linear Regression with R Programming (R-studio)

In this project the we can see how a Simple Linear Regression can be used for prediction for the percentage of an student based on the no. of study hours.

It is implementing using R programming -rmarkdown report.

We will start with simple linear regression involving two variables.

Problem Statement
In this we need to predict the percentage of marks based on the number of hours they study.
Here we are using two variables Hours and Scores for the regression.